versao 201.03 ESTAVEL. Nesta versao a primeira implementacao da nova Topologia Free Rider Slice foi compilada e testada. O cliente esta bem estavel, mas os detalhes maiores estao no bootstrap que podera passar por alteracoes. Foi incluido o ttlOutFree. Agora, se um peer mudar de commum para free rider (vice versa), caso ele seja adcionado pelo mesmo parceiro, sera removido na lista out antiga

+++ src/common/Messages/MessageChannel.hpp
@@ -8,9 +8,9 @@ class MessageChannel : public Message
 	public:
 		MessageChannel(Message* message) : Message(message->GetFirstByte()) {};
 		MessageChannel(uint8_t channelFlag, bool performingPunch, uint16_t extPort, uint32_t idChannel, uint32_t nowTime,
-				uint16_t maxPeerListOutInformed = -1, uint16_t maxPeerListOutInformed_FREE = -1, uint16_t limitUpload = -1);
+				uint16_t maxPeerListOutInformed = -1, uint16_t maxPeerListOutInformed_FREE = -1, int32_t limitUpload = -1);
 		vector<int> GetHeaderValuesDataVector(uint8_t channelFlag, bool performingPunch, uint16_t extPort, uint32_t idChannel, uint32_t nowTime,
-				uint16_t maxPeerListOutInformed = -1, uint16_t maxPeerListOutInformed_FREE = -1, uint16_t limitUpload = -1);
+				uint16_t maxPeerListOutInformed = -1, uint16_t maxPeerListOutInformed_FREE = -1, int32_t limitUpload = -1);
 		vector<uint8_t> GetHeaderValuesSizeVector();
 };
